+++ src/r1cs.rs
@@ -774,8 +774,10 @@ impl<G: Group> RelaxedR1CSSNARK<G> {
     // compute the full satisfying assignment by concatenating W.W, U.u, and U.X
     let mut z = concat(vec![W.W.clone(), vec![U.u], U.X.clone()]);
 
-    let (num_rounds_x, num_rounds_y) =
-      (S.num_cons.log2() as usize, (S.num_vars.log2() + 1) as usize);
+    let (num_rounds_x, num_rounds_y) = (
+      (S.num_cons as f64).log2() as usize,
+      ((S.num_vars as f64).log2() as usize + 1) as usize,
+    );
 
     // outer sum-check
     let tau = (0..num_rounds_x)
@@ -902,8 +904,10 @@ impl<G: Group> RelaxedR1CSSNARK<G> {
     S.append_to_transcript(b"S", &mut transcript);
     U.append_to_transcript(b"U", &mut transcript);
 
-    let (num_rounds_x, num_rounds_y) =
-      (S.num_cons.log2() as usize, (S.num_vars.log2() + 1) as usize);
+    let (num_rounds_x, num_rounds_y) = (
+      (S.num_cons as f64).log2() as usize,
+      ((S.num_vars as f64).log2() as usize + 1) as usize,
+    );
 
     // outer sum-check
     let tau = (0..num_rounds_x)
@@ -961,7 +965,8 @@ impl<G: Group> RelaxedR1CSSNARK<G> {
             .map(|i| (i + 1, U.X[i]))
             .collect::<Vec<(usize, G::Scalar)>>(),
         );
-        SparsePolynomial::new(S.num_vars.log2() as usize, poly_X).evaluate(&r_y[1..].to_vec())
+        SparsePolynomial::new((S.num_vars as f64).log2() as usize, poly_X)
+          .evaluate(&r_y[1..].to_vec())
       };
       (G::Scalar::one() - r_y[0]) * self.eval_W + r_y[0] * eval_X
     };
